+++ src/plugins/observables.js
@@ -1,0 +1,21 @@
+/* @flow */
+import isObservable from 'is-observable'
+import StyleRule from '../rules/StyleRule'
+import type {Rule} from '../types'
+
+export default {
+  onProcessRule(rule: Rule) {
+    if (!(rule instanceof StyleRule)) return
+    const {style} = rule
+    for (const prop in style) {
+      const value = style[prop]
+      if (!isObservable(value)) continue
+      value.subscribe({
+        next: (nextValue) => {
+          // $FlowFixMe
+          rule.prop(prop, nextValue)
+        }
+      })
+    }
+  }
+}